pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cortex"
version = "0.1.0"
description = "Autonomous coding agent orchestrator with built-in Kanban board"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.0",
    "aiosqlite>=0.20",
    "uvicorn>=0.30",
    "fastapi>=0.115",
    "websockets>=13.0",
    "watchdog>=5.0",
    "bcrypt>=4.0",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
cortex = "cortex.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
